requirements.txt
openslide-python
Pillow